produto delete verificacoes feitas

+++ parte3/web/applicationTier/deleteProd.php
@@ -19,21 +19,34 @@
     $ean = $_REQUEST['EAN'];
 
     if ($ean == "") {
-        echo("<p>EAN vazio<p>");
+        echo("<p>[ERRO]EAN vazio<p>");
+        echo("<button onclick='window.history.back()' style='float:left; clear:both'>Voltar</button>");
         return;
     }
 
     try
     {
         $host = "db.ist.utl.pt";
-        $user ="ist426019";
-        $password = "lvng0049";
+        $user ="ist426018";
+        $password = "fcgs5019";
         $dbname = $user;
         $db = new PDO("pgsql:host=$host;dbname=$dbname", $user, $password);
         $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
         $start = "START TRANSACTION;";
         doQuery($start,$db);
+
+        $exists = "SELECT FROM Produto WHERE ean = '$ean';";
+        $result = $db->query($exists);
+
+        if($result->rowCount()==0){
+            $rollback = "ROLLBACK;";
+            doQuery($rollback,$db);
+            echo("<p>[ERRO] Nao existe produto com o ean especificado<p>");
+            echo("<button onclick='window.history.back()' style='float:left; clear:both'>Voltar</button>");
+            return;
+        }
+
         $sql = "DELETE FROM Produto WHERE ean = '$ean';";
         doQuery($sql,$db);
         $end = "COMMIT;";
